requirements.txt
# 基础依赖
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
pillow>=10.0.0
markdown>=3.5.2
olefile>=0.47
rich>=13.7.0

# 内容处理
python-frontmatter>=1.0.0
google-generativeai>=0.3.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0 